Moved Resize bar into separate component
